splitting agents to their own files
