+++ nodes/agent/create_agent.py
@@ -1,0 +1,6 @@
+from .base_agent import BaseAgent
+
+default_prompt = "{{ input_string }}"
+
+
+class CreateAgent(BaseAgent): ...
